% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colley.R
\name{colley_solve_system}
\alias{colley_solve_system}
\title{Solve the Colley linear system}
\usage{
colley_solve_system(C, b)
}
\arguments{
\item{C}{Colley matrix}

\item{b}{numeric vector}
}
\value{
numeric vector of ratings
}
\description{
Internal heper that solves the linear system to produce the rating vector
}
\keyword{internal}
