% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colley.R
\name{colley_team_index}
\alias{colley_team_index}
\title{Create team index for the Colley method}
\usage{
colley_team_index(games)
}
\arguments{
\item{games}{A cleaned games data frame}
}
\value{
A named integer vector
}
\description{
Internal helper that maps team names to integer indices
}
\keyword{internal}
