% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colley.R
\name{fit_colley}
\alias{fit_colley}
\title{Fit Colley ratings}
\usage{
fit_colley(games)
}
\arguments{
\item{games}{A data fram containing at least the columns listed above}
}
\value{
A tibble with columns:
\itemize{
\item \code{team} Team name
\item \code{rating} Colley rating
}
Sorted from highest to lowest rating
}
\description{
Compute team ratings using the Colley matrix method based on win/loss
outcomes in the supplied games data. They Colley system constructs a linear
system \eqn{C r = b}, where \eqn{C} is the Colley matrix and \eqn{b} is a
vector derived from wins and losses by each team in the season.
}
\details{
Required columns in \code{games} are:
\itemize{
\item \code{season}
\item \code{week}
\item \code{home_team}
\item \code{away_team}
\item \code{home_points}
\item \code{away_points}
}
}
\examples{
games <- tibble::tibble(
season = c(2024, 2024),
week = c(1, 2),
home_team = c("Texas A&M", "Alabama"),
away_team = c("Alabama","Texas A&M"),
home_points = c(21, 0),
away_points = c(14, 70)
)

ratings <- fit_colley(games)
}
